Move more code over from retroarch.c to input_driver.c
